requirements.txt
certifi==2018.1.18
cycler==0.10.0
matplotlib==2.1.2
numpy==1.14.0
pandas==0.22.0
pgnparser==1.0
pyparsing==2.2.0
pypng==0.0.18
python-chess==0.22.1
python-dateutil==2.6.1
pytz==2017.3
scipy==1.0.0
seaborn==0.8.1
six==1.11.0
